third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGMaskElement-dom-maskUnits-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'maskUnits' attribute of the SVGMaskElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS maskElement.getAttribute('maskUnits') is "objectBoundingBox"
PASS maskElement.getAttribute('maskUnits') is "userSpaceOnUse"
PASS successfullyParsed is true

TEST COMPLETE
